Moving exception to the right package
