+++ src/main/resources/META-INF/MANIFEST.MF
@@ -2,4 +2,3 @@ Manifest-Version: 1.0
 Implementation-Title: EV3Dev-lang-java
 Implementation-Version: 2.1.0-SNAPSHOT
 Implementation-Vendor: Juan Antonio Breña Moral
-Main-Class: example.LCDDrawRectanglesAnimationTest
